requirements.txt
requests
beautifulsoup4
lxml
html5lib
fake-useragent
urllib3
aiohttp
selenium
webdriver-manager
pandas
numpy
python-dateutil
matplotlib
seaborn
scikit-learn
pyodbc
streamlit
tqdm
python-dotenv
loguru
pip
setuptools
wheel

